# jsonschema: Switch to map based encoding