--- a/schema/encoding/jsonschema/array.go
+++ b/schema/encoding/jsonschema/array.go
@@ -1,28 +1,29 @@
 package jsonschema
 
-import (
-	"io"
-	"strconv"
+import "github.com/rs/rest-layer/schema"
 
-	"github.com/rs/rest-layer/schema"
-)
+type arrayBuilder schema.Array
 
-func encodeArray(w io.Writer, v *schema.Array) error {
-	ew := errWriter{w: w}
-
-	ew.writeString(`"type": "array"`)
+func (v arrayBuilder) BuildJSONSchema() (map[string]interface{}, error) {
+	m := map[string]interface{}{
+		"type": "array",
+	}
 	if v.MinLen > 0 {
-		ew.writeFormat(`, "minItems": %s`, strconv.FormatInt(int64(v.MinLen), 10))
+		m["minItems"] = v.MinLen
 	}
 	if v.MaxLen > 0 {
-		ew.writeFormat(`, "maxItems": %s`, strconv.FormatInt(int64(v.MaxLen), 10))
+		m["maxItems"] = v.MaxLen
 	}
 	if v.ValuesValidator != nil {
-		ew.writeString(`, "items": {`)
-		if ew.err == nil {
-			ew.err = encodeValidator(w, v.ValuesValidator)
+		builder, err := ValidatorBuilder(v.ValuesValidator)
+		if err != nil {
+			return nil, err
 		}
-		ew.writeString("}")
+		items, err := builder.BuildJSONSchema()
+		if err != nil {
+			return nil, err
+		}
+		m["items"] = items
 	}
-	return ew.err
+	return m, nil
 }