--- a/schema/encoding/jsonschema/float.go
+++ b/schema/encoding/jsonschema/float.go
@@ -1,44 +1,32 @@
 package jsonschema
 
 import (
-	"io"
 	"math"
-	"strconv"
-	"strings"
 
 	"github.com/rs/rest-layer/schema"
 )
 
-func encodeFloat(w io.Writer, v *schema.Float) error {
-	ew := errWriter{w: w}
+type floatBuilder schema.Float
 
-	ew.writeString(`"type": "number"`)
+func (v floatBuilder) BuildJSONSchema() (map[string]interface{}, error) {
+	m := map[string]interface{}{
+		"type": "number",
+	}
+
 	if len(v.Allowed) > 0 {
-		var allowed []string
-		for _, value := range v.Allowed {
-			allowed = append(allowed, strconv.FormatFloat(value, 'E', -1, 64))
-		}
-		ew.writeFormat(`, "enum": [%s]`, strings.Join(allowed, ","))
+		m["enum"] = v.Allowed
 	}
-	if ew.err == nil {
-		ew.err = boundariesToJSONSchema(w, v.Boundaries)
+	if v.Boundaries != nil {
+		addBoundariesProperties(m, v.Boundaries)
 	}
-	return ew.err
+	return m, nil
 }
 
-// boundariesToJSONSchema writes JSON Schema keys and values based on b, prefixed by a comma and without curly braces,
-// to w. The prefixing comma is only written if at least one key/value pair is also written.
-func boundariesToJSONSchema(w io.Writer, b *schema.Boundaries) error {
-	if b == nil {
-		return nil
-	}
-	ew := errWriter{w: w}
-
+func addBoundariesProperties(m map[string]interface{}, b *schema.Boundaries) {
 	if !math.IsNaN(b.Min) && !math.IsInf(b.Min, -1) {
-		ew.writeFormat(`, "minimum": %s`, strconv.FormatFloat(b.Min, 'E', -1, 64))
+		m["minimum"] = b.Min
 	}
 	if !math.IsNaN(b.Max) && !math.IsInf(b.Max, 1) {
-		ew.writeFormat(`, "maximum": %s`, strconv.FormatFloat(b.Max, 'E', -1, 64))
+		m["maximum"] = b.Max
 	}
-	return ew.err
 }