--- a/schema/encoding/jsonschema/integer.go
+++ b/schema/encoding/jsonschema/integer.go
@@ -1,26 +1,19 @@
 package jsonschema
 
-import (
-	"io"
-	"strconv"
-	"strings"
+import "github.com/rs/rest-layer/schema"
 
-	"github.com/rs/rest-layer/schema"
-)
+type integerBuilder schema.Integer
 
-func encodeInteger(w io.Writer, v *schema.Integer) error {
-	ew := errWriter{w: w}
+func (v integerBuilder) BuildJSONSchema() (map[string]interface{}, error) {
+	m := map[string]interface{}{
+		"type": "integer",
+	}
 
-	ew.writeString(`"type": "integer"`)
 	if len(v.Allowed) > 0 {
-		var allowed []string
-		for _, value := range v.Allowed {
-			allowed = append(allowed, strconv.FormatInt(int64(value), 10))
-		}
-		ew.writeFormat(`, "enum": [%s]`, strings.Join(allowed, ","))
+		m["enum"] = v.Allowed
 	}
-	if ew.err == nil {
-		ew.err = boundariesToJSONSchema(w, v.Boundaries)
+	if v.Boundaries != nil {
+		addBoundariesProperties(m, v.Boundaries)
 	}
-	return ew.err
+	return m, nil
 }