--- a/schema/encoding/jsonschema/object.go
+++ b/schema/encoding/jsonschema/object.go
@@ -1,15 +1,28 @@
 package jsonschema
 
 import (
-	"io"
+	"fmt"
 
 	"github.com/rs/rest-layer/schema"
 )
 
-func encodeObject(w io.Writer, v *schema.Object) error {
-	ew := errWriter{w: w}
-	if ew.err == nil && v.Schema != nil {
-		ew.err = encodeSchema(w, v.Schema)
+type objectBuilder schema.Object
+
+var (
+	//ErrNoSchema is returned when trying to JSON Encode a schema.Object with the Schema property set to nil.
+	ErrNoSchema = fmt.Errorf("no schema defined for object")
+)
+
+func (v objectBuilder) BuildJSONSchema() (map[string]interface{}, error) {
+	if v.Schema == nil {
+		return nil, ErrNoSchema
 	}
-	return ew.err
+
+	m := map[string]interface{}{}
+	err := addSchemaProperties(m, v.Schema)
+	if err != nil {
+		return nil, err
+	}
+	return m, nil
+
 }