--- a/schema/encoding/jsonschema/string.go
+++ b/schema/encoding/jsonschema/string.go
@@ -1,32 +1,25 @@
 package jsonschema
 
-import (
-	"fmt"
-	"io"
-	"strconv"
-	"strings"
+import "github.com/rs/rest-layer/schema"
 
-	"github.com/rs/rest-layer/schema"
-)
+type stringBuilder schema.String
 
-func encodeString(w io.Writer, v *schema.String) error {
-	ew := errWriter{w: w}
-	ew.writeString(`"type": "string"`)
+func (v stringBuilder) BuildJSONSchema() (map[string]interface{}, error) {
+	m := map[string]interface{}{
+		"type": "string",
+	}
+
 	if v.Regexp != "" {
-		ew.writeFormat(`, "pattern": %q`, v.Regexp)
+		m["pattern"] = v.Regexp
 	}
 	if len(v.Allowed) > 0 {
-		var allowed []string
-		for _, value := range v.Allowed {
-			allowed = append(allowed, fmt.Sprintf("%q", value))
-		}
-		ew.writeFormat(`, "enum": [%s]`, strings.Join(allowed, ", "))
+		m["enum"] = v.Allowed
 	}
 	if v.MinLen > 0 {
-		ew.writeFormat(`, "minLength": %s`, strconv.FormatInt(int64(v.MinLen), 10))
+		m["minLength"] = v.MinLen
 	}
 	if v.MaxLen > 0 {
-		ew.writeFormat(`, "maxLength": %s`, strconv.FormatInt(int64(v.MaxLen), 10))
+		m["maxLength"] = v.MaxLen
 	}
-	return ew.err
+	return m, nil
 }