@smyrman 2017-01-01
@rs, @yanfali. This provides the following public interface for custom validators to implement:

```
+type Formatter interface {
 +	// JSONSchema should return a map containing JSON Schema Draft 4 properties that can be set based on
 +	// FieldValidator data. Application specific properties can be added as well, but should not conflict with any
 +	// legal JSON Schema keys.
 +	JSONSchema() (map[string]interface{}, error)
  }
```

This is, based on "Solution 2" from #35, even you preferred solution 1 there @rs. The internal logic is also switched to rely on maps.

As far as I can see, this solution has the following benefits over solution 1:
- Simpler
- Allows future support for users to encode JSON Schema into other formats than JSON, e.g. YAML as supported by Swagger.
- Avoids a corner case for an io.Writer based solution to #35 where no bytes are written by a custom validator. See [schema.go#L104](https://github.com/smyrman/rest-layer/blob/c671d40ef98b4a37f133f22d8e395377213dd1c7/schema/encoding/jsonschema/schema.go#L104)

It has the following disadvantages compared to solution 1:
- Slower