@yanfali 2017-01-01
Maybe builder then. Formatter suggests final output

On Sun, Jan 1, 2017, 12:38 Sindre Røkenes Myren <notifications@github.com>
wrote:

> *@smyrman* commented on this pull request.
> ------------------------------
>
> In schema/encoding/jsonschema/schema.go
> <https://github.com/rs/rest-layer/pull/76>:
>
> >
> -func (fw *fieldWriter) resetPropertiesCount() {
> -	fw.propertiesCount = 0
> -}
> +	for key, field := range fields {
> +		if field.Required {
> +			required = append(required, key)
> +		}
> +		formatter, err := validatorFormatter(field.Validator)
>
> I call it formatter in line with the ResponseForamtter interface
> <http://rest-layer.io/#custom-response-formatter-sender> from rest-layer.
> It's not an encoder anymore, as it doesn't encode the data, only format the
> map which will be encoded later.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/rs/rest-layer/pull/76>, or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ACK5U7SVMTPyTReFLkNlx_xwkvioSDt1ks5rOA7CgaJpZM4LYt8C>
> .
>