@yanfali 2017-01-01
I find setting null in json tends to be worthless. Much better to not set
the field at all and use undefined in JS or OK pattern in go.

I still think for nil it's much better to set the type for consistency at
the receiver when emitting as it's explicit. Implicit behavior tends to get
people doing different things because of interpretation. I'm arguing set
the type to object for the nil case.

On Sun, Jan 1, 2017, 12:50 Sindre Røkenes Myren <notifications@github.com>
wrote:

> *@smyrman* commented on this pull request.
> ------------------------------
>
> In schema/encoding/jsonschema/example_test.go
> <https://github.com/rs/rest-layer/pull/76>:
>
> >  	// | | },
>  	// | | "foo": {
> -	// | | | "type": "number",
> -	// | | | "minimum": 0E+00
> +	// | | | "minimum": 0,
> +	// | | | "type": "number"
>  	// | | },
>  	// | | "foobar": {}
>
> ah, foobar. That is what we should encode if the field.Validator is set to
> nil. We could change the example, but I think it's the correct think to
> write in this situation. If it's set to nil, we can't enforce it to be an
> object.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/rs/rest-layer/pull/76>, or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ACK5U9QgZIL3c94Ky0kD9YRgOx4GblWWks5rOBGogaJpZM4LYt8C>
> .
>