@smyrman 2017-01-01
> I find setting null in json tends to be worthless. Much better to not set
the field at all and use undefined in JS or OK pattern in go.

True, but that is besides the point...

> I still think for nil it's much better to set the type for consistency at
the receiver when emitting as it's explicit. Implicit behavior tends to get
people doing different things because of interpretation. I'm arguing set
the type to object for the nil case.

Maybe you misunderstand?

I meen if `schema.Field.Validator` is not set or equal to `nil` (in Go), we don't know which _type_ the value may have. The way I read the code in the `schema` package, it may have _any_ type. So it's pretty simple what the `jsonschema` package should do.

The behaviour as it stands is identical to what it was before this change.

The only viable alternative would be to raise an error if we don't want to support Validaor set to nil in the `rest-layer/schema` package, but I personally don't see why we would want that.