@smyrman 2017-01-01
On another topic, the last change addresses most of the review comments.

I will do one more commit soon where I remove the size hints. and rerun the benchmarks. I will leave it as one commit for now and rebase later, in case you want to add/runn one more benchmarks @yanfali.